refactor: combined input/output files
