--- include/file_input.h
@@ -1,8 +1,0 @@
-#ifndef FILE_INPUT_H
-#define FILE_INPUT_H
-
-#include "graph.h"
-
-int ReadInputFile(const char *FilePath, GraphType *graph);
-
-#endif
