+++ include/file_io.h
@@ -1,12 +1,13 @@
-#ifndef FILE_OUTPUT_H
-#define FILE_OUTPUT_H
+#ifndef FILE_IO_H
+#define FILE_IO_H
 
 #include "graph.h"
 
-// ADD FILE NAME PARAMETERS + VERTICES AND EDGES IF NECESSARY
+int ReadInputFile(const char *FilePath, GraphType *graph);
+
 void PrintSet(GraphType graph);
 void PrintDegree(GraphType graph);
 void PrintList(GraphType graph);
 void PrintMatrix(GraphType graph);
 
-#endif
+#endif // FILE_IO_H
